requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
playwright
sentence-transformers
scikit-learn
numpy
aiofiles
httpx
python-dotenv
